Move presence system away from Bot into BotBase
