+++ src/base.ts
@@ -1,12 +1,14 @@
 import autobind from "autobind-decorator";
-import { Client, ClientOptions, Intents } from "discord.js";
-import { report as _report } from "./utility";
+import { ActivityOptions, Client, ClientOptions, Intents } from "discord.js";
+import { report as _report, pick } from "./utility";
 
 let botCount = 0;
+type PresenceType = Required<ActivityOptions["type"]>;
 
 export class BotBase {
   protected token: string;
   id: number;
+  private presenceInterval: NodeJS.Timeout;
 
   /**Discord.js client object */
   readonly client: Client;
@@ -28,6 +30,30 @@ export class BotBase {
         `Bot created and logged in => (${this.client.user?.tag ?? "NO TAG!!!"})`
       );
     });
+  }
+
+  setPresence(
+    activities: [string, PresenceType] | [string, PresenceType][],
+    interval: number = 10 * 60 * 1000 /*10 minutes*/
+  ) {
+    function setActivity(this: BotBase, activity: [string, PresenceType]) {
+      this.client.user?.setActivity(activity[0], { type: activity[1] }) ??
+        this.report(
+          "User missing from client object, bot was unable to update presence."
+        );
+    }
+
+    if (activities.length === 0)
+      throw new Error("Presence list can't be empty");
+
+    clearInterval(this.presenceInterval);
+    if (activities[0] instanceof Array) {
+      this.presenceInterval = setInterval(() => {
+        setActivity.bind(this, pick(activities))();
+      }, interval);
+    } else {
+      setActivity.bind(this, activities)();
+    }
   }
 
   @autobind
